+++ src/async-iterable.d.ts
@@ -1,2 +1,0 @@
-import { asyncIterable } from './index'
-export default asyncIterable