+++ src/iterable.d.ts
@@ -1,2 +1,0 @@
-import { iterable } from './index'
-export default iterable